port a fix from Saurabh

+++ apf_sim/apfSIM.cc
@@ -565,7 +565,7 @@ class IntTagSIM : public TagSIM
 {
   public:
     IntTagSIM(pParMesh m, const char* name, int c):
-      TagSIM(m,name,sizeof(double),c)
+      TagSIM(m,name,sizeof(int),c)
     {
       MD_setMeshCallback(id,CBdelete,deleteIntCB,NULL);
       MD_setMeshCallback(id,CBmigrateOut,pm_sendIntArray,comm);
